fix: Add keys to list elements
Add keys to elements in the metadata and citations that are structured
in a list.  Note I had to replace a couple instances of `<>` fragment
syntax in favor of the explicit `<Fragment` to pass in a key prop.

This fixes *most* warnings for missing / duplicate keys - there are some
edge cases I've seen that are a bit tricky for me, and would have to
take more investigation