Push collection children pagination into next api
Instead of doing all the pagination client side when fetching a
collection's children, fetch it all during the next API
`collectionchildren` request.  This reduces the number of roundtrips
between the client and server. I noticed that even when the server
caches its `fetch` requests, the sheer number of requests from the
client makes this whole thing slow.