Forward client IP through fetch
Currently, our server looks for the IP of the user in various headers to
determine rights for the user. However, since our fetch requests are all
placed on the nextJS server, living in EC2, the IPs for all the requests
end up being EC2 IPs, so our rights logic fails.

To fix this, use `next/headers` to inspect the headers of the incoming
request and forward the origin IP to the API. I used the following
nextJS doc as a guide for this:
https://nextjs.org/docs/14/app/api-reference/functions/headers#ip-address

I kinda need to see what this looks like when it goes through the
reverse proxy, so once this is in, I'll inspect the logs and patch on
the server side.