Redirect items/{captureuuid} to item url with idx
OG DC supports passing in a capture uuid to the items endpoint,
displaying the specific capture on the items page.  On facelift, we
currently also support this type of url, but don't link to the specific
capture, just the item.

To support old links, check if the given uuid is a capture, and if so,
redirect to the url for its parent item uuid, passing in the canvasIndex
corresponding to the specific capture.

Note - this doesn't cover the other type of capture url OG DC supports -
/items/{uuid}#/?uuid={captureuuid}. Those urls will need to be handled
differently, since everything after the `#` is chopped off before
sending off to the server, so we'll need a client side solution in a
followup PR