Redirect items/{captureuuid} to item url with idx
OG DC supports passing in a capture uuid to the items endpoint,
displaying the specific capture on the items page.  On facelift, we
currently also support this type of url, but don't link to the specific
capture, just the item.

To support old links, check if the given uuid is a capture, and if so,
redirect to the url for its parent item uuid, passing in the canvasIndex
corresponding to the specific capture.

For hash-style urls, the uuid param gets lopped off by the browser, so
we parse it client side and use the window API to redirect to the
correct canvas index. Note this looks janky - we end up loading the
page, then reloading to get to the correct index, as I had trouble
getting this to work purely client side.

I think a better option going forward would be to make the capture
linking (whether using canvas index or uuid) purely client side. We
could pass down an array of captures to the viewer, and use OG style
hash params to do the manipulation purely client-side, but I'll leave
that to another day.