Track AV progress and send to google analytics
Use the plyr on 'timeupdate' event to track when specific milestones are
reached when playing an AV file. The idea here is to track when a
'milestone' percentage gets crossed, and send an event for those
milestones. I basically followed the same logic as in OG DC:
https://github.com/NYPL/digitalreadingroom/blob/9c01d999ef8085b867869f65d87cbbc2bbaedeca/app/assets/javascripts/ga-plugin.js#L199-L217
where we wait until a given milestone is passed, but if we're more than
5% past the milestone, we don't trigger an event, to account for the
case where a user is clicking around (ie, if the user skips straight to
40% into the video, we don't want to trigger an event for 25% progress
reached).

Note, I had to split the plyr out into a child component to cleanly use
the `usePlyr`, as in this example from the plyr-react library:
https://github.com/chintan9/plyr-react/blob/eb579712a2e3a7b871a24daa7241acf01856699d/example/src/custom-audio-player.tsx#L18-L50

Also note that for the 100& milestone, I simply use the playback `end`
event.