Pass items endpoint data into item model
We want to fully remove these manifest fetches, but in order to do so,
I'll need to add more data to the existing items endpoint. In the
meantime, get the ball rolling by using the items endpoint where I can
and adding an interface to make the data types clearer.