Use items endpoint for manifest metadata
Final piece to stop calling the manifests endpoint from the items page -
get the manifest metadata from the items endpoint and pass it down as we
are currently doing.  Moving forward, we can work to get rid of the
manifest metadata format in favor of passing typed data, but this is an
expedient stopgap to improve perf on item pages with a lot of captures.