Consume fetch response body on error
Doing some digging, I read that a potential source of memory leaks is
not consuming the response body of a fetch response. Noticed that we
don't consume the body on error, so changed things up a bit to parse and
log the text of the body if we don't get an ok response.

Did a bit of refactoring as a part of this, namely, getting rid of the
nested error handling and removing the 422 specific handling, which
probably made sense when this was first written but now is probably fine
to remove since we have a stabilized API.