Explicitly avoid items page cache
Currently, we use `revalidatePath` to wipe the cache from the items page
- reading a bite more, it seems like a simpler approach would be to just
not cache the data used to load the items page, and then avoid calling
revalidate.  In practice, the items endpoint is already not cached
anyway, since nextJS has some underlying magic to not cache any of the
data if 'dynamic' functions like `headers()` are called (we call this
method to fetch the client side IP address).

In practice, this should do nothing, except maybe reduce a tiny bit of
overhead in calling `revalidatePath`. I think this might just make some
of the cache behavior a bit clearer. Additionally, I think the purpose
of `revalidatePath` is to call it when data is mutated, ie, if you have
a form submission or button that triggers data updates, and you want to
wipe the cache so that the next GET isn't stale.  Since DC is read-only,
seems like we can avoid that step.

If this looks okay on QA, I may reevaluate our other uses of
revalidatePath as well.